any ability commit
made a way to give any pokemon an ability with code. had to add new var to the ability info. then ad to make space because of memory size. removed one save slot and reordered memory storage. And i removed parts of contests and pokemon vars such as ribbon bools

+++ src/battle_pike.c
@@ -1110,7 +1110,7 @@ bool32 TryGenerateBattlePikeWildMon(bool8 checkKeenEyeIntimidate)
     u8 headerId = GetBattlePikeWildMonHeaderId();
     u32 lvlMode = gSaveBlock2Ptr->frontier.lvlMode;
     const struct PikeWildMon *const *const wildMons = sWildMons[lvlMode];
-    u32 abilityNum;
+    u16 abilityNum;
     s32 pikeMonId = GetMonData(&gEnemyParty[0], MON_DATA_SPECIES, NULL);
     pikeMonId = SpeciesToPikeMonId(pikeMonId);
